added some tests

+++ src/Tests/test_vecs.cc
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "../Vector2.h"
+#include "../../include/Vector2.h"
 
 TEST(TestVector2i, BasicAssertions)
 {
@@ -45,13 +45,13 @@ TEST(TestVector2f, BasicAssertions)
     EXPECT_FLOAT_EQ(v3.y, 6.f);
 
     auto v4 = v2 - v1;
-    EXPECT_TRUE(v4 == Vector2i(0, 0));
+    EXPECT_TRUE(v4 == Vector2f(0, 0));
 
     auto v5 = v1 * 3;
-    EXPECT_TRUE(v5 == Vector2i(3, 9));
+    EXPECT_TRUE(v5 == Vector2f(3, 9));
 
     v5 = 3 * v1;
-    EXPECT_TRUE(v5 == Vector2i(3, 9));
+    EXPECT_TRUE(v5 == Vector2f(3, 9));
 }
 
 TEST(VecCasting, BasicAssertions)
